fixes #5: Update Hawtio and Jolokia
This replaces the versions of Hawtio and Jolokia with more up-to-date versions.

Note that the Hawtio web application that is now used includes Jolokia, so we need no add that as a seperate dependency to this plugin ourselves. One downside of this is that the endpoint where the Jolokia service is exposed changed, from `/jolokia/` to `/hawtio/jolokia/`

Instead of copying the binary distributions of the third party dependencies, this commit pulls them in through maven.

I've significantly boosted the version number of the plugin, to reflect that it's nearing completion.